refactor: CartHeader & CartFooter added
