+++ layouts/default.vue
@@ -1,3 +1,6 @@
+<script setup lang="ts">
+  import CartSidebar from '~/components/cart/CartSidebar.vue'
+</script>
 <template>
   <div class="app-container">
     <CartSidebar />
